Correct apparent fractional division
Division between two integers needs promotion to float to avoid modular
truncation.  In this case, the denominator was *currently* defined to 1,
so we'd be safe, but if it ever changed, we'd have a hard-to-find bug.

+++ src/handler_ft8.c
@@ -105,7 +105,7 @@ void sendFT8Tone(long prior_frequency, long frequency, TickType_t *lastWakeTime,
 
     for (int step = 1; step <= EASE_STEPS; step++)
     {
-        long eased_frequency = prior_frequency + round(delta_frequency * (step / EASE_STEPS));
+        long eased_frequency = prior_frequency + round(delta_frequency * ((float)step / EASE_STEPS));
         snprintf(command, sizeof(command), "FA%011ld;", eased_frequency);
 
         // Send the tone command over UART
